Fix: 모델 수정 #1
