+++ makourse/account/admin.py
@@ -1,6 +1,8 @@
 from django.contrib import admin
 from .models import *
 
+admin.site.register(CustomUser)
+
 class UserAdmin(admin.ModelAdmin):
     list_display = ('id', 'name')
 admin.site.register(User, UserAdmin)
